Fix: Support stack-based containers in SrcMov
SrcMov moved the container into itself and directly initialized the iterator.
If SrcMov itself (together with the container) was moved somewhere else, the
iterator was essentially invalidated, because it was pointing to a stack-address.
Fixed this by moving the used collection into the heap, instead of storing it
in SrcMov itself. This ensures the memory address stays the same.