Add ContiguousMemoryIterator trait, implement chunkedExactPtr() chainer
Iterators can now "communicate" whether their elements are stored in a contiguous block of memory, via
implementing the ContiguousMemoryIterator trait. This trait grants access to a pointer on the current element.

Based on that, the chunkedExactPtr() method can now be used to avoids copying elements into a temporary buffer (as
is done by the chunkedExact() chainer). As the benchmarke results show, the savings of that start to set in with
overlapping chunks.